Split GGNN tests into multiple files.
Rather than having a single file with over 2000 tests, split into
multiple tests.
