+++ deeplearning/ml4pl/models/ggnn/BUILD
@@ -19,7 +19,10 @@
 py_binary(
     name = "ggnn",
     srcs = ["ggnn.py"],
-    visibility = ["//deeplearning/ml4pl/experiments:__subpackages__"],
+    visibility = [
+        "//deeplearning/ml4pl/experiments:__subpackages__",
+        "//deeplearning/ml4pl/models/ggnn:__subpackages__",
+    ],
     deps = [
         ":ggnn_config",
         ":ggnn_modules",
@@ -36,26 +39,6 @@ py_binary(
         "//labm8/py:progress",
         "//third_party/py/numpy",
         "//third_party/py/torch",
-    ],
-)
-
-py_test(
-    name = "ggnn_test",
-    size = "enormous",
-    timeout = "long",
-    srcs = ["ggnn_test.py"],
-    shard_count = 16,
-    deps = [
-        ":ggnn",
-        ":ggnn_config",
-        "//deeplearning/ml4pl/graphs/labelled:graph_tuple_database",
-        "//deeplearning/ml4pl/models:batch_iterator",
-        "//deeplearning/ml4pl/models:epoch",
-        "//deeplearning/ml4pl/models:log_database",
-        "//deeplearning/ml4pl/models:logger",
-        "//deeplearning/ml4pl/testing:random_graph_tuple_database_generator",
-        "//deeplearning/ml4pl/testing:testing_databases",
-        "//labm8/py:test",
     ],
 )
